+++ TransitionTreasury/TRNavgationTransitionDelegate.swift
@@ -10,25 +10,31 @@ import UIKit
 /// Transition(Push & Pop) Animation Delegate Object
 public class TRNavgationTransitionDelegate: NSObject, UINavigationControllerDelegate {
     /// Control transition precent
-    var percentTransition: UIPercentDrivenInteractiveTransition?
+//    var percentTransition: UIPercentDrivenInteractiveTransition?
     /// The transition animation object
     public var transition: TRViewControllerAnimatedTransitioning
     
     public var previousStatusBarStyle: TRStatusBarStyle?
     
     public var currentStatusBarStyle: TRStatusBarStyle?
+    
+    private var p_completion: (() -> Void)?
+    
     /// Transition completion block
     var completion: (() -> Void)? {
         get {
-            return self.transition.completion
+            return self.transition.completion ?? p_completion
         }
         set {
             self.transition.completion = newValue
+            if self.transition.completion == nil {
+                p_completion = newValue
+            }
         }
     }
     /// The edge gesture for pop
     lazy var edgePanGestureRecognizer: UIScreenEdgePanGestureRecognizer = {
-        let edgePanGestureRecognizer = UIScreenEdgePanGestureRecognizer(target: self, action: Selector("edgePan:"))
+        let edgePanGestureRecognizer = UIScreenEdgePanGestureRecognizer(target: self, action: Selector("tr_edgePan:"))
         edgePanGestureRecognizer.edges = .Left
         return edgePanGestureRecognizer
     }()
@@ -41,10 +47,10 @@ public class TRNavgationTransitionDelegate: NSObject, UINavigationControllerDele
      
      - returns: Transition Animation Delegate Object
      */
-    public init(method: TRPushMethod, status: TransitionStatus = .Push, gestureFor viewController: UIViewController?) {
+    public init(method: TRPushTransitionMethod, status: TransitionStatus = .Push, gestureFor viewController: UIViewController?) {
         transition = method.transitionAnimation()
         super.init()
-        if transition.edgeSlidePop {
+        if let transition = transition as? TransitionInteractiveable where transition.edgeSlidePop {
             viewController?.view.addGestureRecognizer(edgePanGestureRecognizer)
         }
     }
@@ -60,7 +66,10 @@ public class TRNavgationTransitionDelegate: NSObject, UINavigationControllerDele
     }
     
     public func navigationController(navigationController: UINavigationController, interactionControllerForAnimationController animationController: UIViewControllerAnimatedTransitioning) -> UIViewControllerInteractiveTransitioning? {
-        return transition.interacting ? (transition.percentTransition ?? percentTransition) : nil
+        guard let transition = transition as? TransitionInteractiveable else {
+            return nil
+        }
+        return transition.interacting ? transition.percentTransition : nil
     }
     
     public func navigationController(navigationController: UINavigationController, didShowViewController viewController: UIViewController, animated: Bool) {
@@ -74,10 +83,14 @@ public class TRNavgationTransitionDelegate: NSObject, UINavigationControllerDele
         }
     }
     
-    public func edgePan(recognizer: UIPanGestureRecognizer) {
+    public func tr_edgePan(recognizer: UIPanGestureRecognizer) {
         
         let fromVC = transition.transitionContext?.viewControllerForKey(UITransitionContextFromViewControllerKey)
         let toVC = transition.transitionContext?.viewControllerForKey(UITransitionContextToViewControllerKey)
+        
+        guard var transition = transition as? TransitionInteractiveable else {
+            return
+        }
         
         let view = fromVC!.view
         
@@ -88,23 +101,23 @@ public class TRNavgationTransitionDelegate: NSObject, UINavigationControllerDele
         switch recognizer.state {
         case .Began :
             transition.interacting = true
-            percentTransition = UIPercentDrivenInteractiveTransition()
-            percentTransition!.startInteractiveTransition(transition.transitionContext!)
+            transition.percentTransition = UIPercentDrivenInteractiveTransition()
+            transition.percentTransition?.startInteractiveTransition((transition as! TRViewControllerAnimatedTransitioning).transitionContext!)
             toVC!.navigationController!.tr_popViewController()
         case .Changed :
-            percentTransition?.updateInteractiveTransition(percent)
+            transition.percentTransition?.updateInteractiveTransition(percent)
         default :
             transition.interacting = false
             if percent > transition.interactivePrecent {
                 transition.cancelPop = false
-                percentTransition!.completionSpeed = 1.0 - percentTransition!.percentComplete
-                percentTransition?.finishInteractiveTransition()
+                transition.percentTransition?.completionSpeed = 1.0 - transition.percentTransition!.percentComplete
+                transition.percentTransition?.finishInteractiveTransition()
                 fromVC?.view.removeGestureRecognizer(edgePanGestureRecognizer)
             } else {
                 transition.cancelPop = true
-                percentTransition?.cancelInteractiveTransition()
+                transition.percentTransition?.cancelInteractiveTransition()
             }
-            percentTransition = nil
+            transition.percentTransition = nil
         }
     }
 }