+++ TransitionTreasury/TransitionAnimation/PhotoTransitionAnimation.swift
@@ -24,6 +24,8 @@ public class PhotoTransitionAnimation: NSObject, TRViewControllerAnimatedTransit
     
     public var completion: (() -> Void)?
     
+    public var cancelPop: Bool = false
+    
     private var frameBackup: CGRect?
     
     private var gesturesBackup: [UIGestureRecognizer]?
@@ -125,9 +127,7 @@ public class PhotoTransitionAnimation: NSObject, TRViewControllerAnimatedTransit
                 interacting = true
                 percentTransition = UIPercentDrivenInteractiveTransition()
                 percentTransition!.startInteractiveTransition(transitionContext!)
-                if let detailVC = detailVC as? MainViewControllerDelegate {
-                    detailVC.modalDelegate?.modalViewControllerDismiss(callbackData: nil)
-                }
+                showVC?.tr_dismissViewController()
             } else {
                 interacting = false
             }