+++ TransitionTreasuryDemo/MainViewController.swift
@@ -12,16 +12,18 @@ import TransitionTreasury
 struct PushTransition {
     let name: String
     let imageName: String
-    let pushMethod: TRPushMethod
+    let pushMethod: TRPushTransitionMethod
+    let interactive: Bool
 }
 
 struct PresentTransition {
     let name: String
     let imageName: String
-    let presentMethod: TRPresentMethod
+    let presentMethod: TRPresentTransitionMethod
+    let interactive: Bool
 }
 
-class MainViewController: UIViewController, ModalViewControllerDelegate {
+class MainViewController: UIViewController, ModalTransitionDelegate {
     
     var tr_transition: TRViewControllerTransitionDelegate?
     
@@ -49,24 +51,25 @@ class MainViewController: UIViewController, ModalViewControllerDelegate {
     // MARK: - Modal viewController delegate
     
     func modalViewControllerDismiss(callbackData data:AnyObject? = nil) {
-        presentResultLabel.text = "CallbackData: \(data)."
+        presentResultLabel.text = "CallbackData: \(data?["title"] as? String ?? "")."
         tr_dismissViewController()
         
     }
     
     func loadTransition() {
-        pushTransition.append(PushTransition(name: "OmniFocus", imageName: "OmniFocus60x60", pushMethod: .OMIN(keyView: logoImageView)))
-        pushTransition.append(PushTransition(name: "IBanTang", imageName: "IBanTang60x60", pushMethod: .IBanTang(keyView: logoImageView)))
-        pushTransition.append(PushTransition(name: "Fade", imageName: "WeChat60x60", pushMethod: .Fade))
-        pushTransition.append(PushTransition(name: "Page", imageName: "MeituanMovie60x60", pushMethod: .Page))
-        pushTransition.append(PushTransition(name: "Blixt", imageName: "Blixt60x60", pushMethod: .Blixt(keyView: logoImageView, to: CGRect(x: 30, y: 360, width: logoImageView.frame.size.width / 3, height: logoImageView.frame.size.height / 3))))
-//        pushTransition.append(PushTransition(name: "Storehouse", imageName: "Storehouse60x60", pushMethod: .Storehouse(keyView: logoImageView)))
+        pushTransition.append(PushTransition(name: "OmniFocus", imageName: "OmniFocus60x60", pushMethod: .OMIN(keyView: logoImageView), interactive: false))
+        pushTransition.append(PushTransition(name: "IBanTang", imageName: "IBanTang60x60", pushMethod: .IBanTang(keyView: logoImageView), interactive: false))
+        pushTransition.append(PushTransition(name: "Fade", imageName: "WeChat60x60", pushMethod: .Fade, interactive: false))
+        pushTransition.append(PushTransition(name: "Page", imageName: "MeituanMovie60x60", pushMethod: .Page, interactive: false))
+        pushTransition.append(PushTransition(name: "Blixt", imageName: "Blixt60x60", pushMethod: .Blixt(keyView: logoImageView, to: CGRect(x: 30, y: 360, width: logoImageView.frame.size.width / 3, height: logoImageView.frame.size.height / 3)), interactive: false))
+//        pushTransition.append(PushTransition(name: "Storehouse", imageName: "Storehouse60x60", pushMethod: .Storehouse(keyView: logoImageView), interactive: false))
         
-        presentTransition.append(PresentTransition(name: "Twitter", imageName: "Twitter60x60", presentMethod: .Twitter))
-        presentTransition.append(PresentTransition(name: "Fade", imageName: "WeChat60x60", presentMethod: .Fade))
-        presentTransition.append(PresentTransition(name: "PopTip", imageName: "Alipay60x60", presentMethod: .PopTip(visibleHeight: 500)))
-        presentTransition.append(PresentTransition(name: "TaaskyFlip", imageName: "Taasky60x60", presentMethod: .TaaskyFlip(blurEffect: true)))
-        presentTransition.append(PresentTransition(name: "Elevate", imageName: "Elevate60x60", presentMethod: .Elevate(maskView: logoImageView, to: UIScreen.mainScreen().tr_center)))
+        presentTransition.append(PresentTransition(name: "Twitter", imageName: "Twitter60x60", presentMethod: .Twitter, interactive: false))
+        presentTransition.append(PresentTransition(name: "Fade", imageName: "WeChat60x60", presentMethod: .Fade, interactive: false))
+        presentTransition.append(PresentTransition(name: "PopTip", imageName: "Alipay60x60", presentMethod: .PopTip(visibleHeight: 500), interactive: false))
+        presentTransition.append(PresentTransition(name: "TaaskyFlip", imageName: "Taasky60x60", presentMethod: .TaaskyFlip(blurEffect: true), interactive: false))
+        presentTransition.append(PresentTransition(name: "Elevate", imageName: "Elevate60x60", presentMethod: .Elevate(maskView: logoImageView, to: UIScreen.mainScreen().tr_center), interactive: false))
+        presentTransition.append(PresentTransition(name: "Scanbot", imageName: "Scanbot60x60", presentMethod: .Scanbot(present: nil, dismiss: nil), interactive: true))
         
     }
 
@@ -118,15 +121,18 @@ extension MainViewController: UITableViewDataSource, UITableViewDelegate {
     
     func tableView(tableView: UITableView, didSelectRowAtIndexPath indexPath: NSIndexPath) {
         tableView.deselectRowAtIndexPath(indexPath, animated: true)
-        print("Click")
         guard let cell = tableView.cellForRowAtIndexPath(indexPath) else {
-            print("Fuck")
             return }
         switch indexPath.section {
         case 0 :
+            guard pushTransition[indexPath.row].interactive == false else {
+                let vc = UIStoryboard(name: "Main", bundle: nil).instantiateViewControllerWithIdentifier("InteractiveViewController")
+                navigationController?.pushViewController(vc, animated: true)
+                return;
+            }
             let vc = UIStoryboard(name: "Main", bundle: nil).instantiateViewControllerWithIdentifier("SecondViewController") as! SecondViewController
             vc.title = pushTransition[indexPath.row].name
-            let updateTransition: TRPushMethod = {
+            let updateTransition: TRPushTransitionMethod = {
                 switch self.pushTransition[indexPath.row].name {
                 case "OmniFocus" :
                     return .OMIN(keyView: cell)
@@ -144,11 +150,16 @@ extension MainViewController: UITableViewDataSource, UITableViewDelegate {
                 print("Push finished.")
             })
         case 1 :
+            guard presentTransition[indexPath.row].interactive == false else {
+                let vc = UIStoryboard(name: "Main", bundle: nil).instantiateViewControllerWithIdentifier("InteractiveViewController")
+                navigationController?.pushViewController(vc, animated: true)
+                return;
+            }
             let vc = UIStoryboard(name: "Main", bundle: nil).instantiateViewControllerWithIdentifier("ModalViewController") as! ModalViewController
             vc.modalDelegate = self
             vc.title = presentTransition[indexPath.row].name
             let nav = UINavigationController(rootViewController: vc)
-            let updateTransition: TRPresentMethod = {
+            let updateTransition: TRPresentTransitionMethod = {
                 switch self.presentTransition[indexPath.row].name {
                 case "Elevate" :
                     return .Elevate(maskView: cell.imageView!, to: UIScreen.mainScreen().tr_center)
@@ -162,9 +173,5 @@ extension MainViewController: UITableViewDataSource, UITableViewDelegate {
         default :
             print("Nothing happened.")
         }
-    }
-    
-    override func preferredStatusBarStyle() -> UIStatusBarStyle {
-        return .LightContent
     }
 }